Seperated Tests For Services
Need to change tests to fit the new services but just outlined for now
